+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2015-08-17
 - Bump boto3 version to 1.1.1
 - `ebzl instances -n` display Name tag value (and sorts the output by it)
+- respect default profile from `credentials` file
 
 ## 2015-07-29
 - added `ecs` module for creating & running EC2 Container Service tasks